@ Consolidate duplicate WAR tests into shared GlassFishDeployWarTest
Both GlassFishManagedDeployWarTest and GlassFishRestDeployWarTest were
identical after refactoring to extend GlassFishDeploymentTestTemplate.
Kept a single copy in the ...test package.

@
