+++ integration-tests/src/test/java/org/jboss/arquillian/container/glassfish/test/GlassFishDeployWarTest.java
@@ -14,11 +14,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.arquillian.container.glassfish.managed;
+package org.jboss.arquillian.container.glassfish.test;
 
-import org.jboss.arquillian.container.glassfish.test.GlassFishDeploymentTestTemplate;
-import org.jboss.arquillian.container.glassfish.test.Greeter;
-import org.jboss.arquillian.container.glassfish.test.GreeterServlet;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit5.ArquillianExtension;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
@@ -27,14 +24,14 @@ import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 /**
- * Verifies arquillian tests can run in container mode with this REST based container.
+ * Verifies arquillian tests can run in client mode with this REST based container.
  *
  * @author <a href="http://community.jboss.org/people/aslak">Aslak Knutsen</a>
  * @author <a href="http://community.jboss.org/people/LightGuard">Jason Porter</a>
  * @author <a href="http://community.jboss.org/people/dan.j.allen">Dan Allen</a>
  */
 @ExtendWith(ArquillianExtension.class)
-public class GlassFishManagedDeployWarTest extends GlassFishDeploymentTestTemplate {
+public class GlassFishDeployWarTest extends GlassFishDeploymentTestTemplate {
 
     @Deployment(testable = false)
     public static WebArchive getTestArchive() {
